+++ src/FluxFlow.Components.Assertions.Composition/AssertionsCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Assertions.Contracts;
 using FluxFlow.Components.Assertions.Nodes;
 using FluxFlow.Components.Assertions.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 using FluxFlow.Mapping;
 
 namespace FluxFlow.Components.Assertions.Composition;