+++ src/FluxFlow.Components.FileSystem.Composition/FileSystemCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.FileSystem.Contracts;
 using FluxFlow.Components.FileSystem.Nodes;
 using FluxFlow.Components.FileSystem.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 
 namespace FluxFlow.Components.FileSystem.Composition;