+++ src/FluxFlow.Components.Http.Composition/HttpCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Http.Contracts;
 using FluxFlow.Components.Http.Nodes;
 using FluxFlow.Components.Http.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 
 namespace FluxFlow.Components.Http.Composition;