+++ src/FluxFlow.Components.Mapping.Composition/MappingCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Mapping.Contracts;
 using FluxFlow.Components.Mapping.Nodes;
 using FluxFlow.Components.Mapping.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 using FluxFlow.Mapping;
 
 namespace FluxFlow.Components.Mapping.Composition;