+++ src/FluxFlow.Components.Payloads.Composition/PayloadsCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Payloads.Contracts;
 using FluxFlow.Components.Payloads.Nodes;
 using FluxFlow.Components.Payloads.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 
 namespace FluxFlow.Components.Payloads.Composition;