+++ src/FluxFlow.Components.Sessions.Composition/SessionsCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Sessions.Contracts;
 using FluxFlow.Components.Sessions.Nodes;
 using FluxFlow.Components.Sessions.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace FluxFlow.Components.Sessions.Composition;