+++ src/FluxFlow.Components.Sources.Composition/SourcesCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Sources.Contracts;
 using FluxFlow.Components.Sources.Nodes;
 using FluxFlow.Components.Sources.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 
 namespace FluxFlow.Components.Sources.Composition;