+++ src/FluxFlow.Components.State.Composition/StateCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.State.Contracts;
 using FluxFlow.Components.State.Nodes;
 using FluxFlow.Components.State.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 using FluxFlow.Mapping;
 
 namespace FluxFlow.Components.State.Composition;