+++ src/FluxFlow.Components.Storage.Composition/StorageCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Storage.Contracts;
 using FluxFlow.Components.Storage.Nodes;
 using FluxFlow.Components.Storage.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace FluxFlow.Components.Storage.Composition;