+++ src/FluxFlow.Components.Validation.Composition/ValidationCompositionNodeRegistryExtensions.cs
@@ -2,7 +2,6 @@ using FluxFlow.Components.Validation.Contracts;
 using FluxFlow.Components.Validation.Nodes;
 using FluxFlow.Components.Validation.Options;
 using FluxFlow.Composition;
-using FluxFlow.Composition.Hosting;
 
 namespace FluxFlow.Components.Validation.Composition;